requirements.txt
fastapi>=0.115.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.30.0
pandas>=2.2.0
gspread>=6.1.2
google-auth>=2.29.0
python-multipart>=0.0.9
pydantic[email]>=2.7.0
email-validator>=2.1.0
